refactor: 예약 시간 관리 계층화 리팩터링
