+++ docs/README.md
@@ -1,6 +1,6 @@
 ## 리팩토링 목록
 - [ ] TODO 제거
 - [ ] 테스트 기능별 리팩토링
-- [ ] reservation의 id 생성 방식 개선
 - [ ] simpleJdbcInsert 적용해보기 https://hyeon9mak.github.io/easy-insert-with-simplejdbcinsert/
 - [ ] 시간 관리 페이지에서 초 단위 뜨는 오류 수정
+- [ ] 전체적인 리팩토링
