+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -1,0 +1,51 @@
+package roomescape.controller;
+
+import java.net.URI;
+import java.util.List;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.domain.ReservationTime;
+import roomescape.dto.ReservationTimeResponse;
+import roomescape.dto.ReservationTimeRequest;
+import roomescape.service.ReservationTimeService;
+
+@RestController
+@RequestMapping("/times")
+public class ReservationTimeController {
+
+    private final ReservationTimeService reservationTimeService;
+
+    public ReservationTimeController(final ReservationTimeService reservationTimeService) {
+        this.reservationTimeService = reservationTimeService;
+    }
+
+    @PostMapping
+    public ResponseEntity<Long> add(@RequestBody final ReservationTimeRequest reservationTimeRequest) {
+        long id  = reservationTimeService.add(reservationTimeRequest);
+        HttpHeaders httpHeaders = new HttpHeaders();
+        httpHeaders.setLocation(URI.create("/times/" + id));
+        return ResponseEntity.ok().headers(httpHeaders).body(id); // TODO: create, 객체 반환으로 바꾸기
+    }
+
+    @GetMapping
+    @ResponseStatus(HttpStatus.OK)
+    public List<ReservationTimeResponse> findAll() {
+        List<ReservationTime> reservationTimes = reservationTimeService.findAll();
+        return ReservationTimeResponse.fromReservationTimes(reservationTimes);
+    }
+
+    @DeleteMapping("/{id}")
+    @ResponseStatus(HttpStatus.OK)
+    public int delete(@PathVariable(name = "id") final long id) {
+        return reservationTimeService.delete(id);
+    }
+}
