+++ src/main/java/roomescape/dao/ReservationTimeDao.java
@@ -1,0 +1,50 @@
+package roomescape.dao;
+
+import java.sql.PreparedStatement;
+import java.sql.Time;
+import java.util.List;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.stereotype.Repository;
+import roomescape.domain.ReservationTime;
+import roomescape.dto.ReservationTimeRequest;
+
+@Repository
+public class ReservationTimeDao {
+
+    private final JdbcTemplate jdbcTemplate;
+    private final RowMapper<ReservationTime> actorRowMapper = (rs, rowNum) ->
+            new ReservationTime(
+                    rs.getLong("id"),
+                    rs.getTime("start_at").toLocalTime());
+
+    public ReservationTimeDao(final JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    public long add(final ReservationTimeRequest reservationTimeRequest) {
+        String sql = "INSERT INTO reservation_time (start_at) VALUES (?)";
+
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+        jdbcTemplate.update(connection -> {
+            PreparedStatement ps = connection.prepareStatement(sql, new String[]{"id"});
+            ps.setTime(1, Time.valueOf(reservationTimeRequest.startAt()));
+            return ps;
+        }, keyHolder);
+
+        long id = keyHolder.getKey().longValue();
+        return id;
+    }
+
+    public List<ReservationTime> findAll() {
+        String sql = "SELECT * FROM reservation_time";
+        return jdbcTemplate.query(sql, actorRowMapper);
+    }
+
+    public int delete(final long id) {
+        String sql = "DELETE FROM reservation_time WHERE id = ?";
+        return jdbcTemplate.update(sql, id);
+    }
+}
