--- src/main/java/roomescape/dto/TimeRequest.java
@@ -1,6 +1,0 @@
-package roomescape.dto;
-
-import java.time.LocalTime;
-
-public record TimeRequest(LocalTime startAt) {
-}
